requirements.txt
psycopg2-binary==2.9.7
redis==5.0.1